Update Harden-Windows-Security.ps1
Improved Optional Windows Features category

+++ Harden-Windows-Security.ps1
@@ -1294,6 +1294,11 @@ try {
                     else {
                         Write-Host 'Microsoft Defender Application Guard is already enabled' -ForegroundColor Green 
                     }
+
+                }
+
+                # Need to split the commands in 2 scriptblocks so we don't get "program PowerShell.exe failed to run: The filename or extension is too long" error
+                powershell.exe {
                 
                     # Enable Windows Sandbox
                     Write-Host "`nEnabling Windows Sandbox" -ForegroundColor Yellow
@@ -1347,9 +1352,9 @@ try {
                     }
             
                     # Uninstall VBScript that is now uninstallable as an optional features since Windows 11 insider Dev build 25309 - Won't do anything in other builds                      
-                    if (Get-WindowsCapability -Online | Where-Object { $_.Name -like '*VBSCRIPT*' }) {
-                        try {
-                            Write-Host "`nUninstalling VBSCRIPT" -ForegroundColor Yellow
+                    if (Get-WindowsCapability -Online | Where-Object { $_.Name -like '*VBSCRIPT*' }) {                        
+                        try {  
+                            Write-Host "`nUninstalling VBSCRIPT" -ForegroundColor Yellow                          
                             Get-WindowsCapability -Online | Where-Object { $_.Name -like '*VBSCRIPT*' } | Remove-WindowsCapability -Online -ErrorAction Stop
                             # Shows the successful message only if removal process was successful                                                      
                             Write-Host 'VBSCRIPT has been uninstalled' -ForegroundColor Green
@@ -1361,39 +1366,54 @@ try {
                     }     
                 
                     # Uninstall Internet Explorer mode functionality for Edge
-                    try {
-                        Write-Host "`nUninstalling Internet Explorer mode functionality for Edge" -ForegroundColor Yellow
-                        Get-WindowsCapability -Online | Where-Object { $_.Name -like '*Browser.InternetExplorer*' } | Remove-WindowsCapability -Online -ErrorAction Stop
-                        # Shows the successful message only if removal process was successful
-                        Write-Host 'Internet Explorer mode functionality for Edge has been uninstalled' -ForegroundColor Green
+                    Write-Host "`nUninstalling Internet Explorer mode functionality for Edge" -ForegroundColor Yellow
+                    if ((Get-WindowsCapability -Online | Where-Object { $_.Name -like '*Browser.InternetExplorer*' }).state -ne 'NotPresent') {
+                        try {                            
+                            Get-WindowsCapability -Online | Where-Object { $_.Name -like '*Browser.InternetExplorer*' } | Remove-WindowsCapability -Online -ErrorAction Stop
+                            # Shows the successful message only if removal process was successful
+                            Write-Host 'Internet Explorer mode functionality for Edge has been uninstalled' -ForegroundColor Green
+                        }
+                        catch {
+                            # show errors
+                            $_
+                        }
                     }
-                    catch {
-                        # show errors
-                        $_
+                    else {
+                        Write-Host 'Internet Explorer mode functionality for Edge is already uninstalled.' -ForegroundColor Green
                     }
 
-                    # Uninstall WMIC
-                    try {
-                        Write-Host "`nUninstalling WMIC" -ForegroundColor Yellow
-                        Get-WindowsCapability -Online | Where-Object { $_.Name -like '*wmic*' } | Remove-WindowsCapability -Online -ErrorAction Stop
-                        # Shows the successful message only if removal process was successful
-                        Write-Host 'WMIC has been uninstalled' -ForegroundColor Green
+                    # Uninstall WMIC 
+                    Write-Host "`nUninstalling WMIC" -ForegroundColor Yellow
+                    if ((Get-WindowsCapability -Online | Where-Object { $_.Name -like '*wmic*' }).state -ne 'NotPresent') {                   
+                        try {                            
+                            Get-WindowsCapability -Online | Where-Object { $_.Name -like '*wmic*' } | Remove-WindowsCapability -Online -ErrorAction Stop
+                            # Shows the successful message only if removal process was successful
+                            Write-Host 'WMIC has been uninstalled' -ForegroundColor Green
+                        }
+                        catch {
+                            # show error
+                            $_
+                        }
                     }
-                    catch {
-                        # show error
-                        $_
+                    else {
+                        Write-Host 'WMIC is already uninstalled.' -ForegroundColor Green
                     }
 
                     # Uninstall Legacy Notepad
-                    try {
-                        Write-Host "`nUninstalling Legacy Notepad" -ForegroundColor Yellow
-                        Get-WindowsCapability -Online | Where-Object { $_.Name -like '*Microsoft.Windows.Notepad.System*' } | Remove-WindowsCapability -Online -ErrorAction Stop
-                        # Shows the successful message only if removal process was successful
-                        Write-Host 'Legacy Notepad has been uninstalled. The modern multi-tabbed Notepad is unaffected.' -ForegroundColor Green
+                    Write-Host "`nUninstalling Legacy Notepad" -ForegroundColor Yellow
+                    if ((Get-WindowsCapability -Online | Where-Object { $_.Name -like '*Microsoft.Windows.Notepad.System*' }).state -ne 'NotPresent') {
+                        try {                            
+                            Get-WindowsCapability -Online | Where-Object { $_.Name -like '*Microsoft.Windows.Notepad.System*' } | Remove-WindowsCapability -Online -ErrorAction Stop
+                            # Shows the successful message only if removal process was successful
+                            Write-Host 'Legacy Notepad has been uninstalled. The modern multi-tabbed Notepad is unaffected.' -ForegroundColor Green
+                        }
+                        catch {
+                            # show error
+                            $_
+                        }
                     }
-                    catch {
-                        # show error
-                        $_
+                    else {
+                        Write-Host 'Legacy Notepad is already uninstalled.' -ForegroundColor Green
                     }
                 }
